Hyphens and spaces are now displayed instead of input fields.
The on-screen keyboard backspace is temperamental and seems to
work only for words after the first.

After pressing a key on the external keyboard it does
not automatically move to the next input field. pressing
enter on the external keyboard should check the word.

The answer displayed after three attempts displays a :
at the end.